Refactoring: component analysis of metadata read from S3 moved to its own module
